falsk/requirements.txt
absl-py==0.9.0
aniso8601==1.2.0
appdirs==1.4.0
astor==0.8.1
astroid==2.3.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Bootstrap==3.2.0.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
gast==0.2.2
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
isort==4.3.21
itsdangerous==0.24
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==0.23
matplotlib==3.2.1
mccabe==0.6.1
networkx==2.4
ngt==1.10.0
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
packaging==16.8
pandas==1.0.3
Pillow==6.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pylint==2.4.4
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
simplejson==3.6.0
six==1.10.0
SQLAlchemy==1.1.5
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.11.2